Systemd socket activation
This commit introduces support for systemd socket activation. When rqbit
is started via systemd socket activation, it will now ignore the listen
address specified on the command-line and/or in environment variables
and instead listen on the socket passed by systemd.

The provided unit is just an example that will download to
$HOME/Downloads. The user should customize this to their needs.