Introduce "block_in_place" concurrency limit
This is an attempt to make rqbit work better with slow disks.
I'm running it on an OpenWRT router with a USB stick.

Sometimes it just blocks and loses all momentum. The Web UI also isn't responsive at those times.
My theory is that the "block_in_place" calls stall the whole runtime because the disk is slow.
Disk utilization is 100% when this happens, but SOME work could still be happening e.g. Web UI at least.

Also hope that it will help smooth out the download speed as it tanks during these moments. But at least
averaged out across all peers should be better.